Change link in summary to button. Add ExerciseViewModelMock.
ExerciseViewModelMock is supposed to contain some example data, which makes it easier to style each part of the application separately. This is perhaps not the perfect solution. Another option is to refactor ExerciseViewModel to use a model for dealing with data, and then use mock model, for the purpose of styling.